+++ web/src/components/tasks/TaskList.tsx
@@ -8,33 +8,18 @@ import {
   patchTask,
 } from "../../api/tasks";
 import {
-  ClickAwayListener,
-  Grow,
   IconButton,
   List,
-  ListItem,
-  ListItemAvatar,
-  ListItemIcon,
-  ListItemText,
-  MenuItem,
-  MenuList,
   Paper,
-  Popper,
   Skeleton,
   Stack,
   TextField,
   Typography,
 } from "@mui/material";
 import { Button } from "@mui/material";
-import RadioButtonUncheckedIcon from "@mui/icons-material/RadioButtonUnchecked";
-import CheckIcon from "@mui/icons-material/Check";
-import {
-  ChevronRightSharp,
-  Delete,
-  MoreVert,
-  Refresh,
-} from "@mui/icons-material";
-import React, { useMemo, useState } from "react";
+import { ChevronRightSharp, Delete, Refresh } from "@mui/icons-material";
+import { useMemo, useState } from "react";
+import TaskListItem from "./Task";
 interface Props {
   taskList: TaskList;
 }
@@ -86,6 +71,18 @@ const TaskList = ({ taskList }: Props) => {
     const updated = tasks.filter((t) => t.id !== task.id);
     mutateTasks(updated, { revalidate: false });
   };
+  const handleUpdate = async (taskId: string, attrs: Partial<Task>) => {
+    patchTask(taskList.id, taskId, attrs).then();
+    const updated = tasks.map((t) => {
+      if (t.id == taskId) {
+        return { ...t, ...attrs };
+      } else {
+        return t;
+      }
+    });
+
+    mutateTasks(updated, { revalidate: false });
+  };
   return (
     <Paper elevation={1} sx={{ borderRadius: 3, p: 2, width: 350 }}>
       <Stack direction="row" alignItems="center">
@@ -107,11 +104,12 @@ const TaskList = ({ taskList }: Props) => {
         <>
           <List>
             {pendingTasks.map((task) => (
-              <Task
+              <TaskListItem
                 key={task.id}
                 task={task}
                 onToggleTask={onToggleTask}
                 onDeleteTask={() => handleDelete(task)}
+                onUpdateTask={handleUpdate}
               />
             ))}
             {completedTasks.length > 0 && (
@@ -133,10 +131,11 @@ const TaskList = ({ taskList }: Props) => {
             )}
             {showCompleted &&
               completedTasks.map((task) => (
-                <Task
+                <TaskListItem
                   task={task}
                   onDeleteTask={() => handleDelete(task)}
                   onToggleTask={onToggleTask}
+                  onUpdateTask={handleUpdate}
                 />
               ))}
           </List>
@@ -157,131 +156,6 @@ const TaskList = ({ taskList }: Props) => {
         </>
       )}
     </Paper>
-  );
-};
-
-interface TaskProps {
-  task: Task;
-  onToggleTask: (task: Task) => void;
-  // bound
-  onDeleteTask: () => void;
-}
-const Task = ({ task, onToggleTask, onDeleteTask }: TaskProps) => {
-  const [open, setOpen] = React.useState(false);
-  const anchorRef = React.useRef<HTMLButtonElement>(null);
-
-  const handleToggle = () => {
-    setOpen((prevOpen) => !prevOpen);
-  };
-
-  const handleClose = (event: Event | React.SyntheticEvent) => {
-    if (
-      anchorRef.current &&
-      anchorRef.current.contains(event.target as HTMLElement)
-    ) {
-      return;
-    }
-
-    setOpen(false);
-  };
-
-  function handleListKeyDown(event: React.KeyboardEvent) {
-    if (event.key === "Tab") {
-      event.preventDefault();
-      setOpen(false);
-    } else if (event.key === "Escape") {
-      setOpen(false);
-    }
-  }
-  return (
-    <ListItem key={task.id} sx={{ p: 0 }} alignItems="center">
-      <ListItemAvatar>
-        <IconButton color={"primary"} onClick={() => onToggleTask(task)}>
-          {task.status == "needsAction" ? (
-            <RadioButtonUncheckedIcon />
-          ) : (
-            <CheckIcon />
-          )}
-        </IconButton>
-      </ListItemAvatar>
-      <ListItemText className="group">
-        <Stack
-          direction="row"
-          align-items="start"
-          justifyContent="space-between"
-        >
-          <Typography
-            variant="body1"
-            className={task.status === "completed" ? "text-gray-600" : ""}
-            sx={{
-              textDecoration:
-                task.status === "completed" ? "line-through" : null,
-            }}
-          >
-            {task.title}
-          </Typography>
-          <IconButton
-            title="More task options"
-            className="group-hover:opacity-100 opacity-0"
-            size="small"
-            ref={anchorRef}
-            id="composition-button"
-            aria-controls={open ? "composition-menu" : undefined}
-            aria-expanded={open ? "true" : undefined}
-            aria-haspopup="true"
-            onClick={handleToggle}
-          >
-            <MoreVert />
-          </IconButton>
-        </Stack>
-        <div>
-          <Popper
-            open={open}
-            anchorEl={anchorRef.current}
-            role={undefined}
-            placement="bottom-start"
-            transition
-          >
-            {({ TransitionProps, placement }) => (
-              <Grow
-                {...TransitionProps}
-                style={{
-                  transformOrigin:
-                    placement === "bottom-start" ? "left top" : "left bottom",
-                }}
-              >
-                <Paper>
-                  <ClickAwayListener onClickAway={handleClose}>
-                    <MenuList
-                      autoFocusItem={open}
-                      id="composition-menu"
-                      aria-labelledby="composition-button"
-                      onKeyDown={handleListKeyDown}
-                    >
-                      <MenuItem
-                        onClick={(e) => {
-                          onDeleteTask();
-                          handleClose(e);
-                        }}
-                      >
-                        <ListItemIcon>
-                          <Delete />
-                        </ListItemIcon>
-                        <ListItemText>Delete</ListItemText>
-                      </MenuItem>
-                    </MenuList>
-                  </ClickAwayListener>
-                </Paper>
-              </Grow>
-            )}
-          </Popper>
-        </div>
-
-        <Typography variant="body2" className="text-gray-500">
-          {task.notes}
-        </Typography>
-      </ListItemText>
-    </ListItem>
   );
 };
 export default TaskList;
